pyproject.toml
[project]
name = "airflow-data-pipeline"
version = "0.1.0"
description = "ETL pipeline using Apache Airflow and MongoDB"
requires-python = ">=3.10,<3.14"
dependencies = [
    "apache-airflow==3.1.7",
    "apache-airflow-providers-mongo>=4.0.0",
    "pandas>=2.0.3",
    "pymongo>=4.5.0",
    "chardet>=5.0.0",
    "pendulum>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.82.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["plugins", "dags"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
